fibonacci4

+++ 0x02-functions_nested_loops/104-fibonacci.c
@@ -6,16 +6,16 @@
 
 int main(void)
 {
-	unsigned long long int fib1 = 1;
-	unsigned long long int fib2 = 2;
-	unsigned long long int fib3;
+	unsigned long int fib1 = 1;
+	unsigned long int fib2 = 2;
+	unsigned long int fib3;
 	int i;
 
-	printf("%llu, %llu, ", fib1, fib2);
+	printf("%lu, %lu, ", fib1, fib2);
 	for (i = 3; i <= 98; i++)
 	{
 		fib3 = fib1 + fib2;
-		printf("%llu", fib3);
+		printf("%lu", fib3);
 		if (i != 98)
 		{
 			printf(", ");
